- Changing topic-and-skills-dashboard-page controller to component (both html and ts files). - Test the new component until it gets 100% coverage.
